requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
blinker==1.8.2
boto3==1.28.74
botocore==1.31.84
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
cement==2.10.14
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptocode==0.1
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
daphne==4.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.5
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.3.0
django-extensions==3.2.3
django-mathfilters==1.0.0
django-pwa==1.1.0
django-storages==1.14.2
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docutils==0.21.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
exponent-server-sdk==2.0.0
Faker==21.0.0
firebase-admin==6.5.0
flake8==6.1.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
ghp-import==2.1.0
git-filter-repo==2.45.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib_metadata==8.5.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mediapipe==0.10.14
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-pdf-export-plugin==0.5.10
ml-dtypes==0.4.0
msgpack==1.0.7
multidict==6.0.5
numpy==1.26.3
oauthlib==3.2.2
openai==1.47.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
Pillow==10.0.1
pipdeptree==2.23.1
platformdirs==4.3.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydotplus==2.0.2
pydyf==0.11.0
pyflakes==3.1.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.2
pyphen==0.16.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==5.0.1
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rst2pdf==0.102
s3transfer==0.7.0
schedule==1.2.2
scipy==1.14.1
semantic-version==2.10.0
service-identity==23.1.0
setuptools==69.0.3
six==1.16.0
smartypants==2.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sounddevice==0.5.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
sqlparse==0.4.4
termcolor==2.4.0
tinycss2==1.3.0
tqdm==4.66.5
twilio==9.2.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.4
whitenoise==6.7.0
yarl==1.9.4
zipp==3.20.2
zope.interface==6.1
zopfli==0.2.3
